idc_okta-workmail/requirements.txt
botocore>=1.36.2
aws-cdk-lib
cdk-nag
aws-cdk.aws-lambda-python-alpha